chore: setup MongoDB connection with motor and python-decouple, rename project files
