--- db/database.py
@@ -1,31 +1,0 @@
-from collections.abc import Generator
-
-from sqlalchemy import create_engine
-from sqlalchemy.orm import Session, declarative_base, sessionmaker
-
-from core.config import get_settings
-
-settings = get_settings()
-
-# For SQLite in local/dev, allow usage across threads. For other DBs this is ignored.
-connect_args = {"check_same_thread": False} if settings.database_url.startswith("sqlite") else {}
-
-engine = create_engine(
-    settings.database_url,
-    echo=settings.database_echo,
-    pool_pre_ping=True,
-    connect_args=connect_args,
-)
-SessionLocal = sessionmaker(bind=engine, autocommit=False, autoflush=False)
-
-Base = declarative_base()
-
-
-def get_db() -> Generator[Session, None, None]:
-    """Provide a transactional database session for FastAPI dependencies."""
-
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
